Favorite button bug
Some weird things are happening in the detail fragment. I only create
one button in the xml, but two appear. On createView, the bottom button
sets the texts like I request. When you click the button, you get the
correct text on the top button. If you back out, then select the movie
again, you get the correct text on the bottom again. I'm not sure what
is going on.